chore(mobile): Menu re-architecture in progress
