+++ apps/mobile/src/screens/DrawerMenu/index.js
@@ -1,0 +1,3 @@
+import component from './DrawerMenu'
+
+export default component
