--- apps/mobile/src/screens/GroupNavigation/index.js
@@ -1,3 +1,0 @@
-import component from './GroupNavigation'
-
-export default component
